sloppy create content
