+++ src/assets/js/cinemaSelect.js
@@ -1,0 +1,7 @@
+document.addEventListener('DOMContentLoaded', () => {
+    let elems = document.querySelectorAll('select');
+    const options = {
+        classes: 'select select--cinema'
+    }
+    let cinemaSelect = M.FormSelect.init(elems, options);
+});
